realignment

+++ Driver.py
@@ -2,8 +2,14 @@ import sys
 import random
 import os
 import numpy as np
-characters={'A':0,'B':1,'C':2,'D':3,'E':4,'F':5,'G':6,'H':7,'I':8,'J':9,'K':10,'L':11,'M':12,'N':13,'O':14,'P':15,'Q':16,'R':17,'S':18,'T':19,'U':20,'V':21,'W':22,'X':23,'Y':24,'Z':25,'a':26,'b':27,'c':28,'d':29,'e':30,'f':31,'g':32,'h':33,'i':34,'j':35,'k':36,'l':37,'m':38,'n':39,'o':40,'p':41,'q':42,'r':43,'s':44,'t':45,'u':46,'v':47,'w':48,'x':49,'y':50,'z':51,' ':52}
-reversedCharacters=characters.__class__(map(reversed, characters.items()))
+
+characters = {'A': 0, 'B': 1, 'C': 2, 'D': 3, 'E': 4, 'F': 5, 'G': 6, 'H': 7, 'I': 8, 'J': 9, 'K': 10, 'L': 11, 'M': 12,
+              'N': 13, 'O': 14, 'P': 15, 'Q': 16, 'R': 17, 'S': 18, 'T': 19, 'U': 20, 'V': 21, 'W': 22, 'X': 23,
+              'Y': 24, 'Z': 25, 'a': 26, 'b': 27, 'c': 28, 'd': 29, 'e': 30, 'f': 31, 'g': 32, 'h': 33, 'i': 34,
+              'j': 35, 'k': 36, 'l': 37, 'm': 38, 'n': 39, 'o': 40, 'p': 41, 'q': 42, 'r': 43, 's': 44, 't': 45,
+              'u': 46, 'v': 47, 'w': 48, 'x': 49, 'y': 50, 'z': 51, ' ': 52}
+reversedCharacters = characters.__class__(map(reversed, characters.items()))
+
 
 # successful input: 2,3 3
 
@@ -17,7 +23,7 @@ def readFile(fileName):
 # returns 1 if valid -1,-2,-3,-4 if not(each error returns a different number)
 def validateKeySize(keySize, key):
     if (keySize[0].isdigit() == 0) or (
-        keySize[1].isdigit() == 0):  # in case the key size contains non numeric characters
+                keySize[1].isdigit() == 0):  # in case the key size contains non numeric characters
         return -5
     if len(key) % 2 == 1:  # in case the key contains an odd number of characters
         return -1
@@ -41,74 +47,78 @@ def validateBlockSize(blockSize, plainText):
     temp = temp.replace('\n', '')
     if temp.isalpha() == 0:  # in case the message contains non alphabetic characters
         return -3
-    if (len(temp) % (int(blockSize[0]) * int(blockSize[1]))) > 0:  # in case the message cannot be cut up into exact blocks
+    if (len(temp) % (
+        int(blockSize[0]) * int(blockSize[1]))) > 0:  # in case the message cannot be cut up into exact blocks
         return -4
     return 1
 
 
-#turn characters matrix to numbers from character tables to encrypt(done)
+# turn characters matrix to numbers from character tables to encrypt(done)
 def turnChar(matrix):
-    converted=[]
-    matrix=np.array(matrix)
+    converted = []
+    matrix = np.array(matrix)
     for i in range(0, len(matrix)):
         converted.append(characters.__getitem__(matrix[i]))
 
     return converted
-#turn numbers to Encrypted characters but returns it as an array
+
+
+# turn numbers to Encrypted characters but returns it as an array
 def getCypher(matrix):
-    matrix=np.array(matrix)
-    matrix=matrix.flatten()
-    cypher=[]
+    matrix = np.array(matrix)
+    matrix = matrix.flatten()
+    cypher = []
     for i in range(0, len(matrix)):
         cypher.append(reversedCharacters.__getitem__(matrix[i]))
     return cypher
 
 
-#m = turnChar(np.array(["H","E","L","L","O"]))
-#print m
-#m=getCypher(m)
-#print (m)
+# m = turnChar(np.array(["H","E","L","L","O"]))
+# print m
+# m=getCypher(m)
+# print (m)
 
-#to get encrypted array(done)
+# to get encrypted array(done)
 def Encrypt(message, key):
-    key=np.mat(key)
-    message=np.mat(message)
-    key=key.astype(np.float)
-    message=message.astype(np.float)
-    cy=np.dot(key,message)
-    cy%=52
+    key = np.mat(key)
+    message = np.mat(message)
+    key = key.astype(np.float)
+    message = message.astype(np.float)
+    cy = np.dot(key, message)
+    cy %= 52
     return cy
 
-#cc=Encrypt(A,np.mat("(1,44);(2,43)"))
-#print cc
-#function to reshape an array to n x m matrix
-def getMatrix(matrix,N,M):
-    matrix.reshape(N,M)
+
+# cc=Encrypt(A,np.mat("(1,44);(2,43)"))
+# print cc
+# function to reshape an array to n x m matrix
+def getMatrix(matrix, N, M):
+    matrix.reshape(N, M)
     return matrix
 
 
-#return the overall encrypted text
+# return the overall encrypted text
 # N & M are the key size
 # M & B are the block size
 # key is a string  and text is a string
-def getEncryptedText(text,key,B,N,M):
-    key=list(key)
-    text=list(text)
-    text=turnChar(text)
-    key=turnChar(key)
-    text=np.array(text)
-    key=np.array(key)
-    text=text.reshape(M,B)
-    key=key.reshape(N,M)
-    encryptedText=Encrypt(text,key)
-    encryptedText=getCypher(encryptedText)
+def getEncryptedText(text, key, B, N, M):
+    key = list(key)
+    text = list(text)
+    text = turnChar(text)
+    key = turnChar(key)
+    text = np.array(text)
+    key = np.array(key)
+    text = text.reshape(M, B)
+    key = key.reshape(N, M)
+    encryptedText = Encrypt(text, key)
+    encryptedText = getCypher(encryptedText)
     return encryptedText
 
 
-#text = "helloo"
-#key =   "Hias"
-#t=getEncryptedText(text,key,3,2,2)
-#print t
+# text = "helloo"
+# key =   "Hias"
+# t=getEncryptedText(text,key,3,2,2)
+# print t
 
 
 while 1:  # key validity while loop
